Add controllers to routes
